# expression: implement vectorized evaluation for `builtinAddDate*Sig`