PCP #12101

### What problem does this PR solve?
1. use go generate to  implement vectorized evaluation for `builtinAddDate*Sig`
2. fix errors when use `ADDDATE` with querying a table :
```bash
mysql> select * from `interval`;
+-------+------+------+--------------------+
| str   | int  | dec  | real               |
+-------+------+------+--------------------+
| 77.89 |   96 |    5 | 2.1426728997115996 |
| 72.14 |   25 |    5 | 1.0116074468331964 |
| 14.51 |    7 |    6 |  7.730443335527452 |
| 94.09 |   45 |    1 |  9.796964383311114 |
| 1.59  |   23 |    8 | 6.7950140907766965 |
+-------+------+------+--------------------+
5 rows in set (0.00 sec)
mysql> select `interval`.dec, ADDDATE(20190101120000,INTERVAL `interval`.dec YEAR) from `interval`;
ERROR 1105 (HY000): baseBuiltinFunc.evalString() should never be called, please contact the TiDB team for help
mysql> select `interval`.dec, ADDDATE(20190101120000,INTERVAL `interval`.dec SECOND) from `interval`;
ERROR 2006 (HY000): MySQL server has gone away
No connection. Trying to reconnect...
Connection id:    3
Current database: test

ERROR 2013 (HY000): Lost connection to MySQL server during query
```

### What is changed and how it works?
bench output:
```bash
go test -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args "builtinAddDateStringStringSig" "builtinAddDateStringIntSig" "builtinAddDateStringRealSig" "builtinAddDateStringDecimalSig" "builtinAddDateIntStringSig" "builtinAddDateIntIntSig" "builtinAddDateIntRealSig" "builtinAddDateIntDecimalSig"
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateStringStringSig-VecBuiltinFunc-8                   453           2501207 ns/op          465907 B/op      11402 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateStringStringSig-NonVecBuiltinFunc-8                393           3031851 ns/op          455452 B/op      11248 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateStringIntSig-VecBuiltinFunc-8                      591           2124397 ns/op          491148 B/op      10826 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateStringIntSig-NonVecBuiltinFunc-8                   440           2745639 ns/op          486689 B/op      10670 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateStringRealSig-VecBuiltinFunc-8                     526           2334019 ns/op          521222 B/op      12350 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateStringRealSig-NonVecBuiltinFunc-8                  397           2985765 ns/op          510683 B/op      12011 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateStringDecimalSig-VecBuiltinFunc-8                  637           1854759 ns/op          437419 B/op      12213 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateStringDecimalSig-NonVecBuiltinFunc-8               456           2568247 ns/op          721078 B/op      17562 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateIntStringSig-VecBuiltinFunc-8                      579           2072919 ns/op          270601 B/op       5161 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateIntStringSig-NonVecBuiltinFunc-8                   468           2474160 ns/op          267970 B/op       5001 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateIntIntSig-VecBuiltinFunc-8                         679           1782884 ns/op          312503 B/op       4724 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateIntIntSig-NonVecBuiltinFunc-8                      513           2253695 ns/op          307821 B/op       4579 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateIntRealSig-VecBuiltinFunc-8                        619           1890243 ns/op          340223 B/op       6267 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateIntRealSig-NonVecBuiltinFunc-8                     475           2438759 ns/op          331833 B/op       5998 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateIntDecimalSig-VecBuiltinFunc-8                     751           1541227 ns/op          246952 B/op       5871 allocs/op
BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinAddDateIntDecimalSig-NonVecBuiltinFunc-8                  542           2188795 ns/op          512712 B/op      10901 allocs/op
BenchmarkVectorizedBuiltinTimeFunc-8                                                                    1000000000               0.0171 ns/op          0 B/op          0 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      23.269s

```

### Check List
Tests 
 - Unit test -- PASS
 - Integration test -- PASS
 - Manual test (add detailed scripts or steps below) -- PASS
    - test `ADDDATE` with querying a table