@catror 2019-11-26
> It seems `builtinSubDate*Sig` can be templated in your PR by a little modification, would you like to integrate them together?

OK, I will integrate them.