@catror 2019-11-27
@SunRunAway PTAL. I templated all `ADDDATE` and `SUBDATE` functions. And, could this PR get more points 😏?