@qw4990 2019-11-28
> @SunRunAway PTAL. I templated all `ADDDATE` and `SUBDATE` functions. And, could this PR get more points 😏?

Thanks for your contribution, this PR vectorizes 32 functions at a time! I think it is worthy of more scores, so we will discuss this with our PCP committee soon :D