@catror 2019-11-28
/run-all-tests