@qw4990 2019-11-29
@catror Please format this file.
```
[2019-11-29T07:11:56.181Z] gofmt (simplify)
[2019-11-29T07:12:02.729Z] expression/bench_test.go
```