@qw4990 2019-11-29
@catror After discussing, we decide to regard this PCP problem as a medium difficulty problem, so you will get 400 points after it is merged :D