@catror 2019-11-29
@qw4990 That's great.  I can finally challenge more difficult problems.