@catror 2019-11-29
/run-all-tests